fix(layering): move classify, guard, transcribe nodes to src/runtime/graph (#1948, #2049 Slice B)
Move three low-coupling graph node modules from telegram_bot/graph/nodes/
to src/runtime/graph/nodes/ as the second slice of the reverse-layering
cleanup:

- classify.py — regex query classification (6 types + canned responses)
- guard.py — prompt injection detection (21 regex patterns EN+RU)
- transcribe.py — voice-to-text via Whisper API (factory pattern)

Each node now imports get_client/observe from src.observability instead
of the telegram_bot shim. Compatibility shims are left in the old
telegram_bot/graph/nodes/ locations, and graph.py lazy imports now
resolve to the canonical src.runtime.graph.nodes paths.

All 164 affected tests pass; contract layering test confirms no new
src/->telegram_bot reverse imports.
