+++ telegram_bot/graph/graph.py
@@ -75,14 +75,14 @@ def build_graph(
     Returns:
         Compiled StateGraph ready for .ainvoke() — context pre-bound.
     """
+    from src.runtime.graph.nodes.classify import classify_node
+    from src.runtime.graph.nodes.guard import guard_node
+    from src.runtime.graph.nodes.transcribe import make_transcribe_node
     from telegram_bot.graph.nodes.cache import cache_check_node, cache_store_node
-    from telegram_bot.graph.nodes.classify import classify_node
     from telegram_bot.graph.nodes.grade import grade_node
-    from telegram_bot.graph.nodes.guard import guard_node
     from telegram_bot.graph.nodes.rerank import rerank_node
     from telegram_bot.graph.nodes.retrieve import retrieve_node
     from telegram_bot.graph.nodes.rewrite import rewrite_node
-    from telegram_bot.graph.nodes.transcribe import make_transcribe_node
 
     # Build run-scoped dependency context — injected into nodes via Runtime.
     ctx: GraphContext = {
